+++ noodles-bam/src/record/convert.rs
@@ -1,4 +1,4 @@
-use std::{io, str};
+use std::io;
 
 use noodles_sam::{self as sam, AlignmentRecord};
 
@@ -61,14 +61,7 @@ impl Record {
         let mut builder = sam::Record::builder();
 
         if let Some(read_name) = self.read_name() {
-            let sam_read_name = str::from_utf8(read_name)
-                .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))
-                .and_then(|s| {
-                    s.parse()
-                        .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))
-                })?;
-
-            builder = builder.set_read_name(sam_read_name);
+            builder = builder.set_read_name(read_name.clone());
         }
 
         builder = builder.set_flags(self.flags());
@@ -169,12 +162,10 @@ mod tests {
 
     fn build_record() -> Result<Record, Box<dyn std::error::Error>> {
         use sam::record::{
-            cigar::op::Kind, quality_scores::Score, Flags, MappingQuality, Position,
+            cigar::op::Kind, quality_scores::Score, Flags, MappingQuality, Position, ReadName,
         };
 
-        use crate::record::{
-            cigar::Op, sequence::Base, Cigar, Data, QualityScores, ReadName, Sequence,
-        };
+        use crate::record::{cigar::Op, sequence::Base, Cigar, Data, QualityScores, Sequence};
 
         let reference_sequence_id = ReferenceSequenceId::from(1);
 
@@ -186,7 +177,7 @@ mod tests {
             .set_mate_reference_sequence_id(reference_sequence_id)
             .set_mate_position(Position::try_from(61153)?)
             .set_template_length(166)
-            .set_read_name(ReadName::try_from(b"r0".to_vec())?)
+            .set_read_name(ReadName::try_new("r0")?)
             .set_cigar(Cigar::from(vec![Op::new(Kind::Match, 4)?]))
             .set_sequence(Sequence::from(vec![Base::A, Base::T, Base::G, Base::C]))
             .set_quality_scores(QualityScores::from(vec![