+++ noodles-bam/src/writer/record.rs
@@ -5,7 +5,7 @@ use noodles_sam::{self as sam, AlignmentRecord};
 
 use super::sam_record::NULL_QUALITY_SCORE;
 use crate::{
-    record::{Cigar, ReadName, ReferenceSequenceId},
+    record::{Cigar, ReferenceSequenceId},
     Record,
 };
 
@@ -122,7 +122,7 @@ where
     writer.write_i32::<LittleEndian>(pos)
 }
 
-fn write_l_read_name<W>(writer: &mut W, read_name: Option<&ReadName>) -> io::Result<()>
+fn write_l_read_name<W>(writer: &mut W, read_name: Option<&sam::record::ReadName>) -> io::Result<()>
 where
     W: Write,
 {
@@ -130,7 +130,7 @@ where
 
     let mut read_name_len = read_name
         .map(|name| name.len())
-        .unwrap_or(crate::record::read_name::MISSING.len());
+        .unwrap_or(sam::record::read_name::MISSING.len());
 
     // + NUL terminator
     read_name_len += mem::size_of::<u8>();
@@ -184,16 +184,16 @@ where
     writer.write_i32::<LittleEndian>(template_length)
 }
 
-fn write_read_name<W>(writer: &mut W, read_name: Option<&ReadName>) -> io::Result<()>
+fn write_read_name<W>(writer: &mut W, read_name: Option<&sam::record::ReadName>) -> io::Result<()>
 where
     W: Write,
 {
-    use crate::record::read_name::MISSING;
+    use sam::record::read_name::MISSING;
 
     const NUL: u8 = 0x00;
 
     if let Some(read_name) = read_name {
-        writer.write_all(read_name)?;
+        writer.write_all(read_name.as_ref())?;
     } else {
         writer.write_all(MISSING)?;
     }
@@ -275,7 +275,7 @@ mod tests {
     fn test_write_record_with_all_fields() -> Result<(), Box<dyn std::error::Error>> {
         use sam::record::{
             cigar::op::Kind, data::field::Tag, quality_scores::Score, Flags, MappingQuality,
-            Position,
+            Position, ReadName,
         };
 
         use crate::record::{
@@ -295,7 +295,7 @@ mod tests {
             .set_mate_reference_sequence_id(reference_sequence_id)
             .set_mate_position(Position::try_from(22)?)
             .set_template_length(144)
-            .set_read_name(ReadName::try_from(b"r0".to_vec())?)
+            .set_read_name(ReadName::try_new("r0")?)
             .set_cigar(Cigar::from(vec![
                 Op::new(Kind::Match, 36)?,
                 Op::new(Kind::SoftClip, 8)?,